+++ src/main/java/ro/uvt/info/designpatternslab2023/repository/AuthorRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ro.uvt.info.designpatternslab2023.models.Author;
 @Repository
-public interface AuthorRepository extends JpaRepository<Author,Integer> {
+public interface AuthorRepository extends JpaRepository<Author, Long> {
+
 }
