+++ src/main/java/ro/uvt/info/designpatternslab2023/repository/BookRepository.java
@@ -4,5 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ro.uvt.info.designpatternslab2023.models.Book;
 @Repository
-public interface BookRepository extends JpaRepository<Book, Integer> {
+public interface BookRepository extends JpaRepository<Book,Long> {
+
 }
